Lots of h files so we can compile

+++ libs/ds/linear/include/dynamic_array.h
@@ -51,7 +51,7 @@ ZDynamicArrayInsert
 (
     ZDynamicArray*      Array,
     char*               Input,
-    int                 Index
+    uint64_t            Index
 );
 
 /*
